Path Tracking and Building Completed
+ Path tracking is functional
+ Path building is functional (but technically flawed) - initially path
building occurred alongside tracking but I realized coding back-tracking
later on would be a nightmare so I separated them.

- Path building is flawed - the algorithm should consolidate movement
along points in a cardinal direction (ex: (0.0 to 0.2) into ONE moveTo
action.  However each square-to-square move is a separate action (ex:
(0,0) to (0,2) is TWO actions.  I've been fighting the algorithm enough
that I'll accept this for now and refactor it later in favor of making
forward progress.